Chat message CRUD Functionalities Done

+++ hangin/base/views.py
@@ -2,7 +2,7 @@ from django.shortcuts import render
 from django.db.models import Q
 from django.http import HttpResponse
 from django.contrib.auth.decorators import login_required
-from .models import Room, Topic
+from .models import Room, Topic, Message
 from .forms import RoomForm
 from django.contrib.auth import authenticate, login, logout
 from django.contrib.auth.models import User
@@ -21,7 +21,7 @@ def loginPage(request):
     if request.user.is_authenticated:
         return redirect('home')
     if request.method == 'POST':
-        username = request.POST.get('username')
+        username = request.POST.get('username').lower() 
         password = request.POST.get('password')
         try:
             user = User.objects.get(username=username)
@@ -60,14 +60,37 @@ def logoutUser(request):
     return redirect('home')
 
 def registerPage(request):
-    page ='register'
     form = UserCreationForm()
+
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request,user)
+            return redirect('home')
+        else:
+            messages.error(request,'An Error Occured during Registration')
     return render(request,'base/login_register.html',{'form':form})
+
 
 def room(request,pk):
     room = Room.objects.get(id = pk)
+    room_messages = room.message_set.all().order_by('-created')
+    participants = room.participants.all()
+
+    if request.method == 'POST':
+        message = Message.objects.create(
+            user = request.user,
+            room = room,
+            body = request.POST.get('body')
+
+        )
+        room.participants.add(request.user)
+        return redirect('room', pk=room.id)
     
-    context = {'room':room}
+    context = {'room':room, 'room_messages':room_messages, 'participants' : participants}
     return render(request,'base/room.html',context)
 
 @login_required(login_url='login')
@@ -107,7 +130,23 @@ def updateRoom(request, pk):
 @login_required(login_url='login')
 def deleteRoom(request,pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse("You are not allowed Here!!")
+    
     if request.method == 'POST':
         room.delete()
         return redirect('home')
     return render(request, 'base/delete.html', {'obj':room})
+
+
+@login_required(login_url='login')
+def deleteMessage(request,pk):
+ 
+    message = Message.objects.get(id=pk)
+    if request.user != message.user :
+        return HttpResponse("You are not allowed Here!!")
+    if request.method == 'POST':
+        message.delete()
+        return redirect('')
+    return render(request, 'base/delete.html', {'obj':message})
